fix: move error to 404.tsx and fix some styles
